requirements_prod.txt
httpx
pylint
ipdb
numpy
pandas
scikit-learn
tqdm
colorama
lightgbm

#GOOGLE
google-cloud-bigquery
google-cloud-storage
pyarrow
db-dtypes

# Geospatial packages
shapely
pyproj
geojson

# API
fastapi
uvicorn

# GCP
google-cloud-bigquery
google-cloud-storage

# Setup
setuptools
